Extract 'MessagePush' and employ singletons
