+++ services/galley/package.yaml
@@ -75,6 +75,7 @@ library:
   - resourcet >=1.1
   - retry >=0.5
   - safe-exceptions >=0.1
+  - semigroups
   - servant
   - servant-client-core
   - servant-server
